requirements.txt
Flask
Flask-SocketIO
python-dotenv
PyJWT
eventlet
google-auth
google-auth-oauthlib
google-auth-httplib2
requests